requirements.txt
streamlit
pandas
matplotlib
seaborn
numpy
wordcloud
pillow
scipy
urlextract
emoji